ENH: Remove schemas and all related code
