--- .dockerignore
@@ -1,6 +1,0 @@
-# Ignore everything
-*
-# Except these
-!nginx.conf
-!tools/modify-schema-files.sh
-!schemas/**
