+++ .github/workflows/ci-fmudataio.yml
@@ -52,11 +52,3 @@ jobs:
             fi
           fi
 
-  docker_build:
-    name: Build Docker image
-    runs-on: [ubuntu-latest]
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Test Docker image can build
-        run: docker build .
