--- .github/workflows/schemas-up-to-date.yml
@@ -1,45 +1,0 @@
-name: Schema up to date
-
-on:
-  pull_request:
-    branches: [main, staging]
-  schedule:
-    - cron: "0 0 * * *"
-
-permissions:
-  contents: read
-  pull-requests: write
-
-jobs:
-  check_schema:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.11"
-
-      - name: Install dependencies
-        run: |
-          pip install pip -U
-          pip install -e .
-
-      - name: Check schema
-        run: |
-          if [[ "${{ github.event.pull_request.base.ref }}" == "staging" ]]; then
-            ./tools/update-schemas --diff --prod
-          else
-            ./tools/update-schemas --diff
-          fi
-
-      - name: Ensure schema validates with AJV
-        run: |
-          npm install ajv ajv-formats
-          ./tools/schema-validate-ajv.mjs ./schemas/
-
-      - name: Ensure examples validates with AJV
-        run: |
-          npm install yaml
-          ./tools/examples-validate-ajv.mjs ./examples/example_metadata ./schemas/
