--- nginx.conf
@@ -1,25 +1,0 @@
-map $arg_json $autoindex {
-    true     @json;
-    default  @html;
-}
-
-server {
-       listen 8080;
-
-       add_header Access-Control-Allow-Origin: *.radix.equinor.com;
-
-       root /app;
-       location /schemas/ {
-              try_files "" $autoindex;
-       }
-
-       location @html {
-              autoindex on;
-              autoindex_format html;
-       }
-
-       location @json { # return json format if uri includes ?json=true
-              autoindex on;
-              autoindex_format json;
-       }
-}
