--- radixconfig.yaml
@@ -1,58 +1,0 @@
-# yaml-language-server: $schema=https://raw.githubusercontent.com/equinor/radix-operator/release/json-schema/radixapplication.json
-# Only works if the redhat yaml ext. in vs-code is installed.
-apiVersion: radix.equinor.com/v1
-kind: RadixApplication
-metadata:
-    name: fmu-schemas
-spec:
-    environments:
-        - name: dev
-          build:
-              from: main
-        - name: staging
-          build:
-              from: staging
-        - name: prod
-    components:
-        - name: main
-          src: "."
-          publicPort: http
-          ports:
-              - name: http
-                port: 8080
-          ingressConfiguration:
-              - stickysessions
-          environmentConfig:
-              - environment: dev
-                horizontalScaling:
-                    minReplicas: 1
-                    maxReplicas: 2
-                resources:
-                    requests:
-                        memory: "1Gi"
-                        cpu: "500m"
-                    limits:
-                        memory: "2Gi"
-                        cpu: "1000m"
-              - environment: staging
-                horizontalScaling:
-                    minReplicas: 1
-                    maxReplicas: 2
-                resources:
-                    requests:
-                        memory: "1Gi"
-                        cpu: "500m"
-                    limits:
-                        memory: "2Gi"
-                        cpu: "1000m"
-              - environment: prod
-                horizontalScaling:
-                    minReplicas: 1
-                    maxReplicas: 2
-                resources:
-                    requests:
-                        memory: "1Gi"
-                        cpu: "500m"
-                    limits:
-                        memory: "2Gi"
-                        cpu: "1000m"
