--- schemas/file_formats/0.1.0/inplace_volumes.json
@@ -1,153 +1,0 @@
-{
-  "$defs": {
-    "Fluid": {
-      "description": "Fluid types used as values in the FLUID column.",
-      "enum": [
-        "oil",
-        "gas",
-        "water"
-      ],
-      "title": "Fluid",
-      "type": "string"
-    },
-    "InplaceVolumesResultRow": {
-      "description": "Represents the columns of a row in a static inplace volumes export.\n\nThese fields are the current agreed upon standard result. Changes to the fields or\ntheir validation should cause the version defined in the standard result schema to\nincrease the version number in a way that corresponds to the schema versioning\nspecification (i.e. they are a patch, minor, or major change).",
-      "properties": {
-        "ASSOCIATEDGAS": {
-          "anyOf": [
-            {
-              "minimum": 0.0,
-              "type": "number"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Associatedgas"
-        },
-        "ASSOCIATEDOIL": {
-          "anyOf": [
-            {
-              "minimum": 0.0,
-              "type": "number"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Associatedoil"
-        },
-        "BULK": {
-          "minimum": 0.0,
-          "title": "Bulk",
-          "type": "number"
-        },
-        "FACIES": {
-          "anyOf": [
-            {
-              "type": "string"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Facies"
-        },
-        "FLUID": {
-          "$ref": "#/$defs/Fluid"
-        },
-        "GIIP": {
-          "anyOf": [
-            {
-              "minimum": 0.0,
-              "type": "number"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Giip"
-        },
-        "HCPV": {
-          "anyOf": [
-            {
-              "minimum": 0.0,
-              "type": "number"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Hcpv"
-        },
-        "LICENSE": {
-          "anyOf": [
-            {
-              "type": "string"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "License"
-        },
-        "NET": {
-          "minimum": 0.0,
-          "title": "Net",
-          "type": "number"
-        },
-        "PORV": {
-          "minimum": 0.0,
-          "title": "Porv",
-          "type": "number"
-        },
-        "REGION": {
-          "title": "Region",
-          "type": "string"
-        },
-        "STOIIP": {
-          "anyOf": [
-            {
-              "minimum": 0.0,
-              "type": "number"
-            },
-            {
-              "type": "null"
-            }
-          ],
-          "default": null,
-          "title": "Stoiip"
-        },
-        "ZONE": {
-          "title": "Zone",
-          "type": "string"
-        }
-      },
-      "required": [
-        "FLUID",
-        "ZONE",
-        "REGION",
-        "BULK",
-        "NET",
-        "PORV"
-      ],
-      "title": "InplaceVolumesResultRow",
-      "type": "object"
-    }
-  },
-  "$id": "https://main-fmu-schemas-dev.radix.equinor.com/schemas/file_formats/0.1.0/inplace_volumes.json",
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "description": "Represents the resultant static inplace volumes parquet file, which is naturally\na list of rows.\n\nConsumers who retrieve this parquet file must read it into a json-dictionary\nequivalent format to validate it against the schema.",
-  "items": {
-    "$ref": "#/$defs/InplaceVolumesResultRow"
-  },
-  "title": "InplaceVolumesResult",
-  "type": "array",
-  "version": "0.1.0"
-}
